+++ src/main/java/com/yzgeneration/evc/domain/member/implement/MemberAppender.java
@@ -1,6 +1,6 @@
 package com.yzgeneration.evc.domain.member.implement;
 
-import com.yzgeneration.evc.common.service.port.RandomHolder;
+import com.yzgeneration.evc.common.implement.port.RandomHolder;
 import com.yzgeneration.evc.domain.member.enums.MemberRole;
 import com.yzgeneration.evc.domain.member.enums.MemberStatus;
 import com.yzgeneration.evc.domain.member.model.Member;
