+++ src/main/java/com/yzgeneration/evc/domain/member/implement/MemberValidator.java
@@ -1,13 +1,13 @@
 package com.yzgeneration.evc.domain.member.implement;
 
-import com.yzgeneration.evc.common.exception.CustomException;
+import com.yzgeneration.evc.exception.CustomException;
 import com.yzgeneration.evc.domain.member.model.MemberAuthenticationInformation;
 import com.yzgeneration.evc.domain.member.model.MemberPrivateInformation;
 import com.yzgeneration.evc.domain.member.service.port.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
-import static com.yzgeneration.evc.common.exception.ErrorCode.*;
+import static com.yzgeneration.evc.exception.ErrorCode.*;
 
 @Component
 @RequiredArgsConstructor
