+++ src/main/java/com/yzgeneration/evc/security/JwtAuthenticationFilter.java
@@ -2,8 +2,8 @@ package com.yzgeneration.evc.security;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.yzgeneration.evc.domain.member.authentication.implement.TokenProvider;
-import com.yzgeneration.evc.common.exception.ErrorCode;
-import com.yzgeneration.evc.common.exception.ErrorResponse;
+import com.yzgeneration.evc.exception.ErrorCode;
+import com.yzgeneration.evc.exception.ErrorResponse;
 import com.yzgeneration.evc.domain.member.service.port.MemberRepository;
 import io.jsonwebtoken.ExpiredJwtException;
 import jakarta.servlet.FilterChain;
